chore: export generated types through tsr-types
