+++ packages/timeline-state-resolver/scripts/schema-types.mjs
@@ -8,23 +8,26 @@ import PrettierConf from '@sofie-automation/code-standard-preset/.prettierrc.jso
  * and auto-generates types for those schemas
  *
  * @todo: auto generate the interfaces.ts in /src as well
- * @todo: expose the generated interfaces in the TSR types package
  *
  ***********************************************************/
 
-const paths = (await fs.readdir('./src/integrations', { withFileTypes: true }))
+const dirs = (await fs.readdir('./src/integrations', { withFileTypes: true }))
 	.filter((c) => c.isDirectory())
-	.map((d) => './src/integrations/' + d.name)
+	.map((d) => d.name)
+const capitalise = (s) => s.slice(0, 1).toUpperCase() + s.slice(1)
+const BANNER =
+	'/**\n * This file was automatically generated by json-schema-to-typescript.\n * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,\n * and run "yarn generate-schema-types" to regenerate this file.\n */'
+let indexFile = BANNER
 
-for (const path of paths) {
+for (const dir of dirs) {
+	const path = './src/integrations/' + dir
 	try {
 		if (!(await fs.stat(path + '/$schemas'))) continue
 	} catch {
 		continue
 	}
 
-	let output =
-		'/* tslint:disable */\n/**\n * This file was automatically generated by json-schema-to-typescript.\n * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,\n * and run "yarn generate-schema-types" to regenerate this file.\n */'
+	let output = BANNER
 
 	try {
 		// compile options from file
@@ -50,7 +53,6 @@ for (const path of paths) {
 			output += '\n' + mappingTypes
 		}
 	} catch (e) {
-		console.error(e)
 		console.error('Error while generating mappings for ' + path + ', continuing...')
 	}
 
@@ -70,18 +72,26 @@ for (const path of paths) {
 			output += '\n' + actionTypes
 		}
 	} catch (e) {
-		console.error(e)
 		console.error('Error while generating actions for ' + path + ', continuing...')
 	}
 
 	// very crude way to create an enum for the actionIds:
-	let actionEnum = 'export enum Actions {\n'
+	let actionEnum = 'export enum ' + capitalise(dir) + 'Actions {\n'
 	for (const id of actionIds) {
-		actionEnum += '\t' + id.slice(0, 1).toUpperCase() + id.slice(1) + " = '" + id + "',\n"
+		actionEnum += '\t' + capitalise(id) + " = '" + id + "',\n"
 	}
 	actionEnum += '}\n'
 
 	output += '\n' + actionEnum
 
-	fs.writeFile(path + '/interfaces.ts', output)
+	fs.writeFile('../timeline-state-resolver-types/src/generated/' + dir + '.ts', output)
+	fs.writeFile(
+		path + '/interfaces.ts',
+		`${BANNER}\nexport * from 'timeline-state-resolver-types/dist/generated/${dir}'\n`
+	)
+	console.log('Wrote ' + dir + '.ts')
+
+	indexFile += `\nexport * from './${dir}'`
 }
+
+fs.writeFile('../timeline-state-resolver-types/src/generated/index.ts', indexFile + '\n')
