+++ packages/timeline-state-resolver/src/integrations/atem/interfaces.ts
@@ -1,9 +1,6 @@
-/* tslint:disable */
 /**
  * This file was automatically generated by json-schema-to-typescript.
  * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
  * and run "yarn generate-schema-types" to regenerate this file.
  */
-export enum Actions {
-	Resync = 'resync',
-}
+export * from 'timeline-state-resolver-types/dist/generated/atem'
