--- packages/timeline-state-resolver/src/integrations/casparCG/$schemas/mappings.json
@@ -1,21 +1,0 @@
-[
-	{
-		"$schema": "https://json-schema.org/draft/2020-12/schema",
-		"$id": "https://example.com/product.schema.json",
-		"title": "CasparCG Mapping",
-		"type": "object",
-		"properties": {
-			"channel": {
-				"type": "number"
-			},
-			"layer": {
-				"type": "number"
-			},
-			"previewWhenNotOnAir": {
-				"type": "boolean"
-			}
-		},
-		"required": ["channel", "layer"],
-		"additionalProperties": false
-	}
-]
