--- packages/timeline-state-resolver/src/integrations/casparCG/$schemas/options.json
@@ -1,27 +1,0 @@
-{
-	"$schema": "https://json-schema.org/draft/2020-12/schema",
-	"$id": "https://example.com/product.schema.json",
-	"title": "CasparCG Options",
-	"type": "object",
-	"properties": {
-		"host": {
-			"type": "string"
-		},
-		"port": {
-			"type": "number"
-		},
-		"retryInterval": {
-			"type": "boolean"
-		},
-		"fps": {
-			"type": "number"
-		},
-		"launcherHost": {
-			"type": "string"
-		},
-		"launcherPort": {
-			"type": "number"
-		}
-	},
-	"additionalProperties": false
-}
